+++ examples/kit-nextjs-article-starter/src/components/breadcrumbs/Breadcrumbs.tsx
@@ -1,5 +1,4 @@
 import type React from 'react';
-import { BreadcrumbsPage, BreadcrumbsProps } from '@/components/breadcrumbs/breadcrumbs.props';
 import {
   Breadcrumb,
   BreadcrumbItem,
@@ -9,6 +8,32 @@ import {
   BreadcrumbSeparator,
 } from '@/components/ui/breadcrumb';
 import { NoDataFallback } from '@/utils/NoDataFallback';
+import { LinkFieldValue } from '@sitecore-content-sdk/nextjs';
+import { ComponentProps } from '@/lib/component-props';
+import { GqlFieldString } from '@/types/gql.props';
+
+/**
+ * Model used for Sitecore Component integration
+ */
+type BreadcrumbsProps = ComponentProps & BreadcrumbsData;
+
+type BreadcrumbsData = {
+  fields: {
+    data: {
+      datasource: {
+        ancestors: BreadcrumbsPage[];
+        name: string;
+      };
+    };
+  };
+};
+
+type BreadcrumbsPage = {
+  name: string;
+  title: GqlFieldString;
+  navigationTitle: GqlFieldString;
+  url?: LinkFieldValue;
+};
 
 export const Default: React.FC<BreadcrumbsProps> = (props) => {
   const { fields } = props;
