+++ examples/kit-nextjs-article-starter/src/components/footer-navigation-callout/FooterNavigationCallout.dev.tsx
@@ -1,8 +1,17 @@
 import React from 'react';
-import { FooterNavigationCalloutProps } from './footer-navigation-callout.props';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { Text } from '@sitecore-content-sdk/nextjs';
+import { Text, Field, LinkField } from '@sitecore-content-sdk/nextjs';
 import { ButtonBase as Button } from '@/components/button-component/ButtonComponent';
+
+interface FooterNavigationCalloutFields {
+  title?: Field<string>;
+  description?: Field<string>;
+  linkOptional?: LinkField;
+}
+
+interface FooterNavigationCalloutProps {
+  fields: FooterNavigationCalloutFields;
+}
 
 export const Default: React.FC<FooterNavigationCalloutProps> = ({ fields }) => {
   const { title, description, linkOptional } = fields;
