--- examples/kit-nextjs-article-starter/src/components/footer-navigation-callout/footer-navigation-callout.props.ts
@@ -1,11 +1,0 @@
-import { Field, LinkField } from '@sitecore-content-sdk/nextjs';
-
-interface FooterNavigationCalloutFields {
-  title?: Field<string>;
-  description?: Field<string>;
-  linkOptional?: LinkField;
-}
-
-export interface FooterNavigationCalloutProps {
-  fields: FooterNavigationCalloutFields;
-}
